+++ tests/browser/pages/invest/invest_feedback.py
