+++ tests/browser/pages/invest/invest_footer.py
