+++ tests/browser/pages/invest/invest_guide.py
