+++ tests/browser/pages/invest/invest_header.py
