+++ tests/browser/pages/invest/invest_home.py
