+++ tests/browser/pages/invest/invest_industries.py
