+++ tests/browser/pages/invest/invest_industry.py
