+++ tests/browser/pages/invest/invest_pixels.py
